Fix editor font size controls with React Context
Root cause: Multiple isolated useEditorPreferences() instances meant
components didn't share state. Controls in App.jsx updated one instance,
but QueryEditor/MonacoViewer had separate instances that never updated.

Solution: Converted to React Context provider pattern
- Created EditorPreferencesContext with shared state
- Wrapped app in EditorPreferencesProvider in main.jsx
- All components now consume same preferences instance
- useEffect hooks in editors now trigger on preference changes
- editor.updateOptions() applies changes immediately

Changes:
- NEW: src/contexts/EditorPreferencesContext.jsx - Context provider
- EDIT: src/hooks/useEditorPreferences.js - Re-export from context
- EDIT: src/main.jsx - Wrap app with provider
- EDIT: src/hooks/useEditorPreferences.test.js - Add provider wrapper
- EDIT: src/App.jsx - Import EditorPreferencesProvider

Tested: Font size changes from 12px to 31px work in real-time
All 13 tests passing
localStorage persistence confirmed working

Fixes the critical UX bug where font size controls appeared to work
but didn't actually change the Monaco editor text size.

Identified by Codex review showing isolated state instances.
